requirements.txt
kivy 
mysql-connector-python 
nltk 
spacy 
tensorflow 
torch 
gtts 
pyttsx3 
SpeechRecognition
